Rename governance -> nns_governnace
