+++ rs/proposals/src/canisters/mod.rs
@@ -1,1 +1,1 @@
-pub mod governance;
+pub mod nns_governance;
